update dao method add, update and delete

+++ src/main/java/com/harry/dao/impl/AbstractDao.java
@@ -46,7 +46,14 @@ public abstract class AbstractDao<T> implements DaoInterface<T> {
     }
 
     @Override
-    public int delete(Integer id) {
-        return jdbcTemplate.update("delete from " + tableName + " where id_=?", id);
+    public boolean delete(Integer id) {
+        boolean result = false;
+        if (this.findOne(id) != null) {
+            int target = jdbcTemplate.update("delete from " + tableName + " where id_=?", id);
+            if (target == 1) {
+                result = true;
+            }
+        }
+        return result;
     }
 }
